legacy/aerocrm-2025/aerocrm_project/asgi.py
import os
from django.core.asgi import get_asgi_application
from dotenv import load_dotenv

load_dotenv()
os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'aerocrm_project.settings')
application = get_asgi_application()



