legacy/aerocrm-2025/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14,<3.15
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
django-cors-headers>=4.0,<5.0
Pillow>=10.0,<11.0
djangorestframework-simplejwt>=5.3,<5.4
celery>=5.3,<5.4
redis>=5.0,<6.0
django-filter>=23.5,<24.0
gunicorn>=21.2.0


